cdk/pyproject.toml
[project]
name = "api_infrastructure"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib==2.147.0", # Use the latest version
    "constructs>=10.0.0,<11.0.0",
]